requirements.txt
torch>=1.7.0
numpy
scipy
opencv-python
matplotlib
torchvision
Pillow 